fix(nil pointer): storage class reference
